requirements.txt
paho-mqtt==1.6.1
opencv-python==4.6.0.66
qrcode==7.4.2
numpy==1.23.0
pycryptodome==3.19.0